refactor(driver-devtool): 组件化项目
