+++ .vscode/settings.json
@@ -2,6 +2,9 @@
     "cSpell.words": [
         "Sval",
         "WECHAT",
+        "Wexin",
+        "Weibo",
+        "Codemirror",
         "wechatsync"
     ],
     "editor.formatOnSave": true
