+++ package.json
@@ -44,6 +44,7 @@
     "style-loader": "^2.0.0",
     "terser-webpack-plugin": "^5.1.1",
     "vue-loader": "^15.9.6",
+    "vue-markdown-loader": "^2.4.1",
     "vue-template-compiler": "^2.6.12",
     "webpack": "^5.25.0",
     "webpack-cli": "^4.5.0",
