+++ packages/driver-devtool/.vscode/settings.json
@@ -1,0 +1,6 @@
+{
+    "editor.formatOnSave": true,
+    "cSpell.words": [
+        "splitpane"
+    ]
+}
