+++ packages/driver-devtool/public/index.html
@@ -9,15 +9,14 @@
         content="微信公众号文章多平台同步，内容营销工具，自媒体内容同步，自媒体助手、内容分发，一键同步发布，支持微博头条、今日头条、豆瓣、WordPress、知乎、简书、掘金、CSDN、typecho各大平台，一次发布，多平台同步发布。解放个人生产力">
     <meta data-n-head="true" name="keywords"
         content="内容营销工具,自媒体助手,自媒体内容同步,内容分发,一键同步发布,微信公众号同步助手,文章同步,独立开发者,WordPress同步,微信文章同步到WordPress,CSDN,博客园,掘金">
-    <link rel="stylesheet" href="assets/fonts/font-awesome/css/font-awesome.min.css" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/highlightjs@9.12.0/styles/github-gist.css">
-    <script src="https://cdn.jsdelivr.net/npm/highlightjs@9.12.0/highlight.pack.min.js"></script>
+    <!-- <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/highlightjs@9.12.0/styles/github-gist.css"> -->
+    <!-- <script src="https://cdn.jsdelivr.net/npm/highlightjs@9.12.0/highlight.pack.min.js"></script> -->
     <!-- <script src="libs/juqery.js"></script> -->
     <!-- <script src="libs/jquery.xmlrpc.min.js"></script> -->
     <!-- <script src="libs/google-analytics-bundle.js"></script> -->
     <!-- wechatsync.com Baidu tongji analytics -->
-    <script>
+    <!-- <script>
         var _hmt = _hmt || [];
         (function () {
             var hm = document.createElement("script");
@@ -34,13 +33,13 @@
                 analytics.page();
             }
         }();
-    </script>
+    </script> -->
 </head>
 
 <body>
-    <main id="app">
-        <router-view />
-    </main>
-    <!-- built files will be auto injected -->
+  <noscript>
+    <strong>We're sorry but the site doesn't work properly without JavaScript enabled. Please enable it to continue.</strong>
+  </noscript>
+  <div id="app"></div>
 </body>
 </html>
