+++ packages/driver-devtool/src/assets/defaultTestArticle.md
@@ -1,0 +1,60 @@
+# # Precibus terrarum confecta misero
+
+## Oras Iuppiter Lernaeae est quoque pectora
+
+Lorem markdownum, cum crimina agrestes: e tolle iungere remolliat melior, Minos
+pariter quos sociis mors. Dominum dis falsam, cum omnis potenti **aetas**;
+mansit modo meritam.
+
+1. His sorores per solet inseris Herculeos
+2. Desertaque metuendus mihi terrae hanc sive vento
+3. Matrumque iam latumque medius quin artes nocent
+4. Quique quod memorque toto iam amara enixa
+
+## Est clipeo Arcade alter
+
+Voce summum Saturnia est quae si latissima nunc damnarat; [et nate
+gaudet](http://precibus.com/) vicinia corripit levius: flebile sum herba. Amante
+Calydonia tetigit, dictaque pondere properant deum centum totiens deus dolusque
+data fert, torquere da! Neque et messis **Palatinae**; restabat interea coniunx!
+
+> Phoebus nec; magno verbis; face tamen. Post terras novantur omnia, est nato
+> sumptaque hanc, vos. Cruentato parcere aliis pectora conata destrinxit quid,
+> intellegat celsior Aeoliden. In esse talia, concursibus armis dixere carina.
+
+## Exaudi gravis materia voto
+
+Deae continet ut demisit vipereos iugales curasque arces resumit alligat elusam
+mollibus Salmacis, fratremque. Voce suo vincla regaliter morientibus feret,
+caeleste torque tumultus navis. In quaeque, tenues ab currebam **piosque
+senumque** visus sanguis motis primos est obnoxia, tectus tecumque munus.
+
+    menu.blobHeaderAnd(wirelessHoverInterface, twitterProgram,
+            latency_troll.schema(binary_blacklist));
+    drag += cableHyperFlops(model_third, ccd);
+    if (bitmap_drive + definition) {
+        server(sector_graymail + 5, dot_day);
+        pci /= nameOfficeMenu;
+    } else {
+        shell_technology += handle - display_interface(337227);
+        server += copyright_cable;
+    }
+
+## Segnis alto
+
+Invasit neci intra homines lapsa ad fidem colloque longior illo, illa, virgine!
+Totumque et matre; utile quod pauca alite haec: aequoreas benefacta. Et tumulo
+*florentis* adfectu, luctus, tenebrasque [monimenta
+corpus](http://non-spectatae.io/) mendacia non telum erat: nec. Pectora
+velamenta male est latrantes pace: ferus hunc festa crines veniente, iam quoque,
+monte mulcet floribus occidit. Orbis nepotis merguntque vidit.
+
+    userNewsgroupLan = batchLeakWidget(hoc_web_header * scroll_text +
+            dialSystemMail);
+    var importUddi = ping_scsi + midi_active;
+    folder -= matrix * desktop + publishing_app_windows;
+
+Perque una propter miscet nisi aut pudorem locavit sperne a. Patetis cornicis
+requiris gratata datis qua. Recentes *me* agmina bella miscet haeserunt aequoris
+**habes penetralia limine**, vana matrem ille qui erit, parere facilem. Cum
+renarro, longe in serpens cessere novo, hoc accessi.
