+++ packages/driver-devtool/src/components/Explorer/CodeEditor.vue
@@ -1,0 +1,10 @@
+<template>
+  <div>Editor</div>
+</template>
+
+<script>
+export default {}
+</script>
+
+<style>
+</style>
