+++ packages/driver-devtool/src/components/Explorer/Console.vue
@@ -1,0 +1,13 @@
+<template>
+
+</template>
+
+<script>
+export default {
+
+}
+</script>
+
+<style>
+
+</style>
