+++ packages/driver-devtool/src/components/Explorer/Statusbar.vue
@@ -1,0 +1,12 @@
+<template>
+  <div class="container"></div>
+</template>
+
+<script>
+export default {}
+</script>
+
+<style>
+.container {
+}
+</style>
