+++ packages/driver-devtool/src/index.js
@@ -1,29 +1,19 @@
 import Vue from 'vue'
-import VueRouter from 'vue-router'
-import VueMoment from 'vue-moment'
 
-//
 import VueCodemirror from 'vue-codemirror'
 import 'codemirror/lib/codemirror.css'
 
-import contentmenu from 'v-contextmenu'
+import contextmenu from 'v-contextmenu'
 import 'v-contextmenu/dist/index.css'
-Vue.use(contentmenu);
-
-import Main from './App.vue'
+Vue.use(contextmenu);
 
 import ElementUI from 'element-ui'
 import 'element-ui/lib/theme-chalk/index.css'
 
 Vue.use(ElementUI)
 
-import mavonEditor from 'mavon-editor'
-import 'mavon-editor/dist/css/index.css'
-import './styles/boot.css'
-import "./main.css";
 
 // or import all icons if you don't care about bundle size
-// import 'vue-awesome/icons'
 /* Register component with one of 2 methods */
 
 // import Icon from 'vue-awesome/components/Icon'
@@ -52,31 +42,7 @@ Vue.directive('highlightjs', {
 });
 
 // use
-Vue.use(mavonEditor)
-Vue.use(VueRouter)
-Vue.use(VueMoment)
 Vue.use(VueCodemirror, /* {
   options: { theme: 'base16-dark', ... },
   events: ['scroll', ...]
 } */)
-
-var routes = [
-  {
-    path: '/',
-    component: Main,
-    meta: {
-      index: 1,
-    },
-  },
-]
-
-// var winBackgroundPage = chrome.extension.getBackgroundPage()
-// var db = winBackgroundPage.db
-// window.db = db
-var router = new VueRouter({
-  routes,
-})
-const app = new Vue({
-  router,
-})
-app.$mount('#app')
