+++ packages/driver-devtool/src/register-directives.js
@@ -1,0 +1,7 @@
+import Vue from 'vue'
+
+Vue.directive('focus', {
+  inserted: function(el) {
+    el.focus()
+  },
+})
