+++ packages/driver-devtool/src/style/global.css
@@ -1,0 +1,45 @@
+
+:root {
+  box-sizing: border-box;
+}
+
+*,
+::before,
+::after {
+  box-sizing: inherit;
+}
+
+:root {
+  --font-primary-color: #1f2329;
+  --font-secondary-color: #5f636d;
+  --line-color: #dee0e3;
+  --hover-secondary-bg-color: #5f636d;
+  --icon-default-color: #5f636d;
+  --foreground-color: #eff3f6;
+  --background-color: #fff;
+  --shadow-color: rgba(0, 0, 0, 0.2);
+
+  --space-width: 0.5em;
+}
+
+
+body {
+  font-family: "Open Sans", -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, "Helvetica Neue", Arial, sans-serif;
+  color: var(--font-primary-color);
+  user-select: none;
+  line-height: 1.3;
+}
+
+* {
+  -webkit-font-smoothing: antialiased;
+  text-rendering: optimizeLegibility;
+  -webkit-tap-highlight-color: rgba(0,0,0,0);
+}
+
+.clearfix::before, .clearfix::after {
+  display: block;
+  content: "";
+}
+.clearfix::after {
+  clear: both;
+}
