+++ packages/driver-devtool/src/style/main.css
@@ -1,0 +1,2 @@
+@import "./normalize.css";
+@import "./global.css";
