+++ packages/driver-devtool/webpack.config.js
@@ -1,19 +1,20 @@
 const path = require('path')
 const webpack = require('webpack')
+const { merge } = require('webpack-merge')
 const HtmlWebpackPlugin = require('html-webpack-plugin')
 const { VueLoaderPlugin } = require('vue-loader')
 const MiniCssExtractPlugin = require('mini-css-extract-plugin')
-const Dotenv = require("dotenv-webpack");
+const Dotenv = require('dotenv-webpack')
 const TerserPlugin = require('terser-webpack-plugin')
 const projectRoot = path.resolve(__dirname, '../../')
 
-module.exports = (env) => {
+module.exports = env => {
   const prodMode = env.production
   const prodConfigs = {
     mode: 'production',
     resolve: {
       alias: {
-        vue: 'vue/dist/vue.min.js'
+        vue: 'vue/dist/vue.min.js',
       },
     },
     optimization: {
@@ -22,7 +23,7 @@ module.exports = (env) => {
         new TerserPlugin({
           test: /\.js(\?.*)?$/i,
         }),
-      ]
+      ],
     },
   }
   const devConfigs = {
@@ -33,65 +34,86 @@ module.exports = (env) => {
     },
     resolve: {
       alias: {
-        vue: 'vue/dist/vue.js'
+        vue: 'vue/dist/vue.js',
       },
     },
   }
-  return {
-    ...(prodMode ? prodConfigs : devConfigs),
-    entry: './src/index.js',
-    output: {
-      filename: 'main.js',
-      path: path.resolve(__dirname, 'dist'),
-      clean: true,
-    },
-    module: {
-      rules: [
-        {
-          test: /\.(sa|sc|c)ss$/i,
-          use: [
-            prodMode ? MiniCssExtractPlugin.loader : 'style-loader',
-            'css-loader'
-          ],
+  return merge(
+    {
+      entry: './src/main.js',
+      output: {
+        filename: 'main.js',
+        path: path.resolve(__dirname, 'dist'),
+        clean: true,
+      },
+      resolve: {
+        alias: {
+          '@': path.resolve(__dirname, 'src'),
         },
-        {
-          test: /\.(png|svg|jpg|jpeg|gif)$/i,
-          type: 'asset/resource',
-        },
-        {
-          test: /\.(woff|woff2|eot|ttf|otf)$/i,
-          type: 'asset/resource',
-        },
-        {
-          test: /\.js?$/,
-          loader: 'babel-loader',
-          exclude: (file) =>
-            /node_modules/.test(file) && !/\.vue\.js/.test(file),
-        },
-        {
-          test: /\.vue$/,
-          loader: 'vue-loader',
-        },
+      },
+      module: {
+        rules: [
+          {
+            test: /\.(sa|sc|c)ss$/i,
+            use: [
+              prodMode ? MiniCssExtractPlugin.loader : 'style-loader',
+              'css-loader',
+              'sass-loader',
+            ],
+          },
+          {
+            test: /\.(png|svg|jpg|jpeg|gif)$/i,
+            type: 'asset/resource',
+          },
+          {
+            test: /\.(woff|woff2|eot|ttf|otf)$/i,
+            type: 'asset/resource',
+          },
+          {
+            test: /\.js?$/,
+            loader: 'babel-loader',
+            exclude: file =>
+              /node_modules/.test(file) &&
+              !/\.vue\.js/.test(file) &&
+              !/node_modules\/vue\-awesome/.test(file),
+          },
+          {
+            test: /\.vue$/,
+            loader: 'vue-loader',
+          },
+          {
+            test: /\.md$/,
+            use: [
+              {
+                loader: 'vue-loader',
+              },
+              {
+                loader: 'vue-markdown-loader',
+              },
+            ],
+          },
+        ],
+      },
+      plugins: [
+        new webpack.ProvidePlugin({
+          Buffer: ['buffer', 'Buffer'],
+        }),
+        new MiniCssExtractPlugin({
+          filename: 'styles.css',
+        }),
+        new HtmlWebpackPlugin({
+          template: 'public/index.html',
+        }),
+        new VueLoaderPlugin(),
+        new Dotenv({
+          path: path.resolve(
+            projectRoot,
+            prodMode ? '.env.production' : '.env.development'
+          ),
+          safe: true,
+        }),
       ],
     },
-    plugins: [
-      new webpack.ProvidePlugin({
-        Buffer: ['buffer', 'Buffer']
-      }),
-      new MiniCssExtractPlugin({
-        filename: 'styles.css'
-      }),
-      new HtmlWebpackPlugin({
-        template: 'src/index.html',
-      }),
-      new VueLoaderPlugin(),
-      new Dotenv({
-        path: path.resolve(
-          projectRoot,
-          prodMode ? '.env.production' : '.env.development'
-        ),
-        safe: true,
-      })
-    ],
-  }
+    prodMode ? prodConfigs : devConfigs
+  )
 }
